Fix merging of adjacent spans
Background: We merge separate spans that cover a continuous logical span of text.
E.g: [(0, 17), (17, 35)] -> [(0, 35)]

Previously we sometimes encountered negative span annotations like this one:
(208, 147). This was caused by incorrect merging of adjacent
annotations. Instead of filtering the end positions using the original
start positions we filtered using the already filtered start positions.
As a result no end positions were ever removed, meaning the function
produced two lists of different lengths, the longer of which (with the
end positions) was truncated by the call to zip.