Allow passing AnnotationCollections to `get_iaa`
Previously this would only work with the names. The old behaviour still
works. The names of annotation collectinos are not guaranteed to be unique.
This meant that previously some collections could not be compared to
others.

This is a general issue with the ac_dict and this patch can only be seen
as a stopgap measure. I am sure related issues will come up in other parts
of the codebase.